docker/utils_LEGO_ml_vision/requirements-utils.txt
# Machine Learning & Deep Learning
ultralytics==8.2.99

# Image Processing
opencv-python==4.9.0.80
pillow==11.1.0
matplotlib==3.8.4

# Data Handling & Utilities
numpy==1.26.4
pandas==2.2.1
pyyaml==6.0.1
json5==0.9.14
glob2==0.7

# CLI & Logging
argparse==1.4.0
tqdm==4.66.2
